types: Move Instantiator, ClaimError and AvmType from core to types
