+++ core/src/display_object/bitmap.rs
@@ -9,8 +9,9 @@ use crate::backend::render::BitmapHandle;
 use crate::context::{RenderContext, UpdateContext};
 use crate::display_object::{DisplayObjectBase, DisplayObjectPtr, TDisplayObject};
 use crate::prelude::*;
-use crate::vminterface::{AvmType, Instantiator};
 use gc_arena::{Collect, Gc, GcCell, MutationContext};
+use ruffle_types::vminterface::AvmType;
+use ruffle_types::vminterface::Instantiator;
 use std::cell::{Ref, RefMut};
 
 /// A Bitmap display object is a raw bitamp on the stage.
