+++ core/src/display_object/movie_clip.rs
@@ -33,9 +33,11 @@ use crate::events::{ButtonKeyCode, ClipEvent, ClipEventResult};
 use crate::font::Font;
 use crate::prelude::*;
 use crate::tag_utils::{self, DecodeResult, SwfMovie, SwfSlice, SwfStream};
-use crate::vminterface::{AvmObject, AvmType, Instantiator};
+use crate::vminterface::AvmObject;
 use gc_arena::{Collect, Gc, GcCell, MutationContext};
 use ruffle_types::string::{AvmString, WStr, WString};
+use ruffle_types::vminterface::AvmType;
+use ruffle_types::vminterface::Instantiator;
 use smallvec::SmallVec;
 use std::cell::{Ref, RefMut};
 use std::collections::HashMap;
