+++ core/src/library.rs
@@ -6,9 +6,9 @@ use crate::display_object::{Bitmap, Graphic, MorphShape, TDisplayObject, Text};
 use crate::font::{Font, FontDescriptor};
 use crate::prelude::*;
 use crate::tag_utils::SwfMovie;
-use crate::vminterface::AvmType;
 use gc_arena::{Collect, MutationContext};
 use ruffle_types::string::AvmString;
+use ruffle_types::vminterface::AvmType;
 use std::collections::HashMap;
 use std::sync::{Arc, Weak};
 use swf::CharacterId;
