+++ core/src/player.rs
@@ -33,12 +33,13 @@ use crate::prelude::*;
 use crate::tag_utils::SwfMovie;
 use crate::timer::Timers;
 use crate::transform::TransformStack;
-use crate::vminterface::{AvmType, Instantiator};
 use gc_arena::{make_arena, ArenaParameters, Collect, GcCell};
 use instant::Instant;
 use log::info;
 use rand::{rngs::SmallRng, SeedableRng};
 use ruffle_types::string::AvmString;
+use ruffle_types::vminterface::AvmType;
+use ruffle_types::vminterface::Instantiator;
 use std::collections::{HashMap, VecDeque};
 use std::ops::DerefMut;
 use std::str::FromStr;
