+++ types/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod numbers;
 pub mod string;
+pub mod vminterface;
